Rename packages as well.
